+++ inst/unitTests/runit.DistributionFits.R
@@ -49,8 +49,11 @@ test.nFit <- function()
     par(mfrow = c(1, 1))
      
     # Simulate normal random variates:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+
+    ## GNB: was
+    ##    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ##    set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     s = rnorm(n = 2000, mean = 1, sd = 0.5) 
     
     # Fit:
@@ -79,8 +82,10 @@ function()
     par(mfrow = c(1, 1))
     
     # Simulate random variates t(4):
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## GNB:
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     s = rt(n = 2000, df = 4)
     
     # Fit:  
@@ -108,8 +113,9 @@ function()
     par(mfrow = c(1, 1))
     
     # Simulate random variates:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     s = rgh(n = 2000, alpha = 0.8, beta = 0.2, delta = 2, mu = -0.4, 
         lambda = 1) 
     
@@ -143,8 +149,9 @@ function()
     par(mfrow = c(1, 1))
     
     # Simulate normal random variates:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     s = rhyp(n = 2000, alpha = 1.5, beta = 0.8, delta = 0.5, mu = -1) 
     
     # Fit:
@@ -175,8 +182,10 @@ test.nigFit <- function()
     par(mfrow = c(1, 1))
     
     # Simulate normal random variates:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4712)
+    
     th.0 <- c(alpha = 1.5, beta = -0.7, delta = 0.5, mu = -1.0)
     s <- do.call(rnig, c(list(n = 2000), th.0))
 
@@ -208,8 +217,9 @@ test.stableFit <- function()
     par(mfrow = c(1, 1))
     
     # Simulate stable random variates:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4713)
 
     th.0 <- c(alpha=1.8, beta=0.3, gamma = 1, delta = 0.1)
     s <- do.call(stabledist::rstable, c(list(n = 500), th.0))
@@ -222,7 +232,7 @@ test.stableFit <- function()
     arelErr <- abs(relErr <- (est - th.0) / th.0)
     print(cbind(th.0, est, relErr))
     checkTrue(all(arelErr[c("alpha","gamma")] < 0.05) &&
-              arelErr[["beta"]] < 0.15 && arelErr[["delta"]] < 1.50) # 150% !
+              arelErr[["beta"]] < 0.20 && arelErr[["delta"]] < 0.50)
 
     # MLE Fit:
     if (FALSE) { # as this takes rather too long time ...
@@ -247,4 +257,3 @@ test.stableFit <- function()
  
    
 ################################################################################
-