+++ inst/unitTests/runit.HyperbolicDistribution.R
@@ -54,8 +54,9 @@ test.gh =
 function()
 {
     # gh() Distribution:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     test = fBasics:::.distCheck("gh",
         alpha = 1.3, beta = 0.3, delta = 1.7, mu = 0.2, lambda = 0.8,
         n = 2000, robust = FALSE)
@@ -74,8 +75,9 @@ test.hyp =
 function()
 {
     # hyp() Distribution - Parameterization 1:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     test = fBasics:::.distCheck("hyp",
         alpha = 1.2, beta = 0.2, delta = 1.9, mu = 0.1, pm = "1",
         n = 1000, robust = FALSE)
@@ -83,8 +85,9 @@ function()
     checkTrue(mean(test) == 1)
 
     # hyp() Distribution - Parameterization 2:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     test = fBasics:::.distCheck("hyp",
         alpha = 0.9, beta = -0.3, delta = 1.4, mu = -0.1, pm = "2",
         n = 1000, robust = FALSE)
@@ -92,8 +95,9 @@ function()
     checkTrue(mean(test) == 1)
 
     # hyp() Distribution - Parameterization 3:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     fBasics:::.distCheck("hyp",
         alpha = 0.9, beta = -0.3, delta = 1.4, mu = -0.1, pm = "3",
         n = 1000, robust = FALSE)
@@ -123,8 +127,9 @@ test.nig =
 function()
 {
     # nig() Distribution:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     test = fBasics:::.distCheck("nig",
         alpha = 2.1, beta = 0.1, delta = 1.5, mu = -0.1,
         n = 1000, robust = FALSE)
@@ -174,4 +179,3 @@ function()
 
 
 ################################################################################
-