+++ inst/unitTests/runit.NormalityTests.R
@@ -54,8 +54,9 @@ test.NormalityTests =
 function()
 {
     # Normal Data:
-    RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
-    set.seed(4711, kind = "Marsaglia-Multicarry")
+    ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
+    ## set.seed(4711, kind = "Marsaglia-Multicarry")
+    set.seed(4711)
     X = rnorm(50)
 
     TEST = ksnormTest(X)
@@ -159,7 +160,4 @@ function()
     return()
 }
 
-
 ################################################################################
-
-